Implement WX report transmission via APRS (CWOP)
This commits adds a new node called "aprs tx" which can be used to
transmit WX reports to APRS via the Citizen Weather Observer Program
(CWOP). CWOP is an extension of APRS-IS which makes it possible to
integrate non-HAM weather stations into APRS. The reports made by such
weather stations are visible, e.g., on aprs.fi, but are not propagated
over the radio (AX.25).

The new node expects a string or object in msg.payload. String payload
is asssumed to be APRS formatted and will be transmitted verbatim.
Object payload is assumed to be encoded in the same format that is
produced by aprs-parse. Object payloads will be serialized and
transmitted to CWOP. Currently, the node can only serialize WX reports.
Other message types are not supported.

Since CWOP servers do not require a passcode, this commit also makes the
pass field in the configuration node optional. The value "-1" will be
used when no passcode is provided.